Add post summarizing FP resources for beginners
Added a post summarizing some basic resources for beginners, describing shortly what the link is about and why it's worth to watch it. Also added myself to the list of the authors.

I added the links and pictures separately, because putting image and it's description in one markdown link did not seem to render correcly. All links are tested, and the article is spell checked.